requirements.txt
streamlit
requests
scikit-learn
hdbscan
pandas
numpy
matplotlib
plotly
python-dotenv
nltk
